Update widgetset for Vaadin 7.2

+++ extras/vaadin/addons/easyuploads/src/main/java/org/vaadin/easyuploads/client/ui/VMultiUpload.java
@@ -264,7 +264,7 @@ public class VMultiUpload extends SimplePanel implements Paintable {
 	 * @return
 	 */
 	private String serialize(VHtml5File file) {
-		int size = file.getSize();
+		double size = file.getSize();
 		String name = file.getName();
 		String type = file.getType();
 		return size + DELIM + name + DELIM + type + DELIM;
